1. 用vs2010编译通过

+++ ixer0.11/ixer/UIClass/SortListCtrl/SortListCtrl.cpp
@@ -363,7 +363,7 @@ void CSortListCtrl::SaveColumnInfo()
 	m_ctlHeader.Serialize( ar );
 	ar.Close();
 
-	DWORD dwLen = memFile.GetLength();
+	DWORD dwLen = (DWORD)memFile.GetLength();
 	BYTE* buf = memFile.Detach();	
 
 	VERIFY( AfxGetApp()->WriteProfileBinary( g_pszSection, strKey, buf, dwLen ) );
